Update until deleteMapping

+++ src/main/java/com/csa/samplefullstack/entity/Employee.java
@@ -15,23 +15,23 @@ public class Employee{
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int employee_id;
+    private Long employee_id;
     private String name;
     private String roles;
     private String nickname;
 
-    public Employee(int employee_id, String name, String roles, String nickname) {
+    public Employee(Long employee_id, String name, String roles, String nickname) {
         this.employee_id = employee_id;
         this.name = name;
         this.roles = roles;
         this.nickname = nickname;
     }
 
-    public int getEmployee_id() {
+    public Long getEmployee_id() {
         return employee_id;
     }
 
-    public void setEmployee_id(int employee_id) {
+    public void setEmployee_id(Long employee_id) {
         this.employee_id = employee_id;
     }
